chore: Refactor decimal validation to the decimal module
